+++ packages/clerk-js/src/globals.d.ts
@@ -1,8 +1,11 @@
 declare global {
   const __DEV__: boolean;
-  const __BUILD_FLAG_ACCOUNTLESS_UI__: boolean;
   const __PKG_NAME__: string;
   const __PKG_VERSION__: string;
+  /**
+   * Build time feature flags.
+   */
+  const __BUILD_FLAG_KEYLESS_UI__: boolean;
   const __BUILD_DISABLE_RHC__: string;
 
   interface Window {