+++ packages/clerk-js/src/ui/contexts/components/Waitlist.ts
@@ -1,12 +1,17 @@
-import { createContext, useContext } from 'react';
+import { createContext, useContext, useMemo } from 'react';
 
 import { buildURL } from '../../../utils';
 import { useEnvironment, useOptions } from '../../contexts';
+import { useRouter } from '../../router';
 import type { WaitlistCtx } from '../../types';
+import { getInitialValuesFromQueryParams } from '../utils';
+
+const WAITLIST_INITIAL_VALUE_KEYS = ['email_address'];
 
 export type WaitlistContextType = WaitlistCtx & {
   signInUrl: string;
   afterJoinWaitlistUrl?: string;
+  initialValues: any;
 };
 
 export const WaitlistContext = createContext<WaitlistCtx | null>(null);
@@ -15,6 +20,12 @@ export const useWaitlistContext = (): WaitlistContextType => {
   const context = useContext(WaitlistContext);
   const { displayConfig } = useEnvironment();
   const options = useOptions();
+  const { queryString } = useRouter();
+
+  const initialValuesFromQueryParams = useMemo(
+    () => getInitialValuesFromQueryParams(queryString, WAITLIST_INITIAL_VALUE_KEYS),
+    [],
+  );
 
   if (!context || context.componentName !== 'Waitlist') {
     throw new Error('Clerk: useWaitlistContext called outside Waitlist.');
@@ -29,5 +40,6 @@ export const useWaitlistContext = (): WaitlistContextType => {
     ...ctx,
     componentName,
     signInUrl,
+    initialValues: { ...initialValuesFromQueryParams },
   };
 };