+++ packages/clerk-js/src/ui/lazyModules/components.ts
@@ -16,8 +16,8 @@ const componentImportPaths = {
   BlankCaptchaModal: () => import(/* webpackChunkName: "blankcaptcha" */ './../components/BlankCaptchaModal'),
   UserVerification: () => import(/* webpackChunkName: "userverification" */ './../components/UserVerification'),
   Waitlist: () => import(/* webpackChunkName: "waitlist" */ './../components/Waitlist'),
-  AccountlessPrompt: __BUILD_FLAG_ACCOUNTLESS_UI__
-    ? () => import(/* webpackChunkName: "accountlessPrompt" */ './../components/AccountlessPrompt')
+  KeylessPrompt: __BUILD_FLAG_KEYLESS_UI__
+    ? () => import(/* webpackChunkName: "keylessPrompt" */ '../components/KeylessPrompt')
     : () => null,
 } as const;
 
@@ -86,9 +86,9 @@ export const BlankCaptchaModal = lazy(() =>
 export const ImpersonationFab = lazy(() =>
   componentImportPaths.ImpersonationFab().then(module => ({ default: module.ImpersonationFab })),
 );
-export const AccountlessPrompt = __BUILD_FLAG_ACCOUNTLESS_UI__
-  ? // @ts-expect-error Types are broken due to __BUILD_FLAG_ACCOUNTLESS_UI__
-    lazy(() => componentImportPaths.AccountlessPrompt().then(module => ({ default: module.AccountlessPrompt })))
+export const KeylessPrompt = __BUILD_FLAG_KEYLESS_UI__
+  ? // @ts-expect-error Types are broken due to __BUILD_FLAG_KEYLESS_UI__
+    lazy(() => componentImportPaths.KeylessPrompt().then(module => ({ default: module.KeylessPrompt })))
   : () => null;
 
 export const preloadComponent = async (component: unknown) => {