+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 2.19.0
+
+### Minor Changes
+
+- Introduce new submodules: ([#4716](https://github.com/clerk/javascript/pull/4716)) by [@panteliselef](https://github.com/panteliselef)
+
+  - Export `OAUTH_PROVIDERS` from `@clerk/shared/oauth`
+  - Export `WEB3_PROVIDERS` from `@clerk/shared/web3`
+
+### Patch Changes
+
+- Updated dependencies [[`0bc3ccc5bd4a93121bb7e7d6a32271af9c31f8c3`](https://github.com/clerk/javascript/commit/0bc3ccc5bd4a93121bb7e7d6a32271af9c31f8c3)]:
+  - @clerk/types@4.38.0
+
+## 2.18.1
+
+### Patch Changes
+
+- Updated dependencies [[`4e5e7f463c12893a21cb3b5f9317fc3f2945879b`](https://github.com/clerk/javascript/commit/4e5e7f463c12893a21cb3b5f9317fc3f2945879b)]:
+  - @clerk/types@4.37.0
+
 ## 2.18.0
 
 ### Minor Changes